Initial commit of working blockToFacet script
This commit introduces the initial version of the blockToFacet glyph
script. This script is designed to take a defined block(s) in a
Pointwise project and directly write out its boundary information to a
Xpatch facet file, including the associated geometry.params file. The
block does not have to be initialized for this script to work and in
fact it is designed to allow us to skip the initialization entirely in
the cases where we only want the boundary definitions from the block.

There is an important caveat: there can only be one 'Undefined' boundary
condition (usually a connection boundary connection) otherwise all of
the undefined boundary conditions are ignored which potentially might
result in a geometry that is not water tight. This issue should not come
up often but might arise if the user is only trying to export a portion
of the blocks in the Pointwise project. Solution: name the connection
boundary conditions.

There is also a little 'easter egg' for the users entertainment.