Remove supposed support for Unspecified BCs
This commit removes the check for an Unspecified boundary condition that
only has one register, this was supposedly a connection boundary
condition. This, however, should never happen and is an incorrect way to
check for an exposed connection. From here on out (for now anyways)
Unspecified boundary conditions are unsupported; the end user will be
warned that Unspecified boundary conditions have been detected and
skipped. All boundaries required for exporting must be specified.

Also, minor output refactoring.